refactor(crypto): move KeyGenerationService callers to legacy-crypto
Point KeyGenerationService and DefaultKeyGenerationService callers at
@bitwarden/legacy-crypto, then delete both shims. The
key-management/crypto barrel has no exports left, so it goes too.

Three of the call sites imported KeyGenerationService without using it;
those imports are dropped rather than migrated. Removes the unused
imports sitting alongside them in the same files, which the pre-commit
eslint hook flags as soon as the files are touched.
