--- libs/common/src/key-management/crypto/index.ts
@@ -1,2 +1,0 @@
-export { KeyGenerationService } from "./key-generation/key-generation.service";
-export { DefaultKeyGenerationService } from "./key-generation/default-key-generation.service";
