--- libs/common/src/key-management/crypto/key-generation/default-key-generation.service.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated Moved to `@bitwarden/legacy-crypto`. These primitives are being retired in favour of
- * the SDK — do not add new callers. Contact the Key Management team.
- */
-export { DefaultKeyGenerationService } from "@bitwarden/legacy-crypto/services/default-key-generation.service";
